test meli y refactor test spotify
