+++ app/src/test/java/bpino/pom/pages/meli/HomePage.java
@@ -1,6 +1,7 @@
 package bpino.pom.pages.meli;
 
 import bpino.pom.base.SeleniumBase;
+import dsaavedra.pom.utilities.ManejoEncodingUFT8;
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
 
@@ -10,17 +11,19 @@ public class HomePage extends SeleniumBase {
     public HomePage(WebDriver driver) {
         super(driver);
     }
-    By locatorTxtCookies = By.xpath("//button[contains(text(),'Entendido')]");
-    By locatorTxtUbicacion = By.xpath(fixEncoding("//span[contains(text(),'Más tarde')]"));
-    By locatorInputProducto = By.xpath("//input[@id='cb1-edit']");
-    By locatorBtnBuscar = By.xpath("//button[@class= 'nav-search-btn']");
 
-    public void buscarProducto(String producto){
-        if(estaDesplegado(locatorTxtCookies)){
-          clickear(locatorTxtCookies);
-          clickear(locatorTxtUbicacion);
-          escribir(producto, locatorInputProducto);
-          clickear(locatorBtnBuscar);
+    By locatorTxtCookies = By.xpath("//button[contains(text(),'Entendido')]");
+    By locatorTxtUbicacion = By.xpath(ManejoEncodingUFT8.fixEncoding("//span[contains(text(),'Más tarde')]"));
+    By locatorInputProducto = By.xpath("//input[@id='cb1-edit']");
+    By locatorBtnBuscar = By.xpath("//button[@class='nav-search-btn']");
+
+
+    public void buscarProducto(String producto) {
+        if (estaDesplegado(locatorTxtCookies)) {
+            clickear(locatorTxtCookies);
+            clickear(locatorTxtUbicacion);
+            escribir(producto, locatorInputProducto);
+            clickear(locatorBtnBuscar);
         }
     }
 }
