+++ app/src/test/java/bpino/pom/pages/meli/LoginPage.java
@@ -1,6 +1,7 @@
 package bpino.pom.pages.meli;
 
 import bpino.pom.base.SeleniumBase;
+import dsaavedra.pom.utilities.ManejoEncodingUFT8;
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
 
@@ -10,7 +11,7 @@ public class LoginPage extends SeleniumBase {
     public LoginPage(WebDriver driver) {
         super(driver);
     }
-    By locatorTxtResultadoEsperado = (By.xpath(fixEncoding("//h1[contains(text(),'¡Hola! Para comprar, ingresa a tu cuenta')]")));
+    By locatorTxtResultadoEsperado = (By.xpath(ManejoEncodingUFT8.fixEncoding("//h1[contains(text(),'¡Hola! Para comprar, ingresa a tu cuenta')]")));
 
     public String textoObtenido(){
         return obtenerTexto(locatorTxtResultadoEsperado);
